refactor: SubwayGraph의 상태에 노선 아이디인 lineId를 넣어 줄 수 있도록 변경
