+++ src/test/java/wooteco/subway/dao/DaoImplTest.java
@@ -1,5 +1,6 @@
 package wooteco.subway.dao;
 
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
