+++ src/test/java/wooteco/subway/domain/SubwayGraphTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.subwaygraph.SubwayGraph;
 import wooteco.subway.exception.CalculatePathsException;
 
 class SubwayGraphTest {
